Intento de Docker compose/backend/consumidor/requirements.txt
requests==2.26.0
confluent-kafka==1.7.0
kafka-python==2.0.2
pyspark==3.3.1
cassandra-driver==3.25.0

